+++ src/NuGet.Core/NuGet.Protocol/Model/ServiceIndexEntry.cs
@@ -2,8 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using NuGet.Configuration;
-using NuGet.Protocol.Events;
 using NuGet.Versioning;
 
 namespace NuGet.Protocol
@@ -28,7 +26,7 @@ namespace NuGet.Protocol
         /// </summary>
         public SemanticVersion ClientVersion { get; }
 
-        internal ServiceIndexEntry(Uri serviceUri, string serviceType, SemanticVersion clientVersion, PackageSource packageSource)
+        public ServiceIndexEntry(Uri serviceUri, string serviceType, SemanticVersion clientVersion)
         {
             if (serviceUri == null)
             {
@@ -45,17 +43,9 @@ namespace NuGet.Protocol
                 throw new ArgumentNullException(nameof(clientVersion));
             }
 
-            if (packageSource != null)
-            {
-                var resourceIsHttp = serviceUri.Scheme == Uri.UriSchemeHttp && serviceUri.Scheme != Uri.UriSchemeHttps;
-                ProtocolDiagnostics.RaiseEvent(new ProtocolDiagnosticServiceIndexEntryEvent(packageSource.Source, resourceIsHttp && packageSource.IsHttps));
-            }
-
             Uri = serviceUri;
             Type = serviceType;
             ClientVersion = clientVersion;
         }
-
-        public ServiceIndexEntry(Uri serviceUri, string serviceType, SemanticVersion clientVersion) : this(serviceUri, serviceType, clientVersion, null) { }
     }
 }
